+++ Game/Source/Container.cpp
@@ -1,6 +1,6 @@
 #include "Container.h"
 
-Container::Container(const long ID, const unsigned int itemType, const unsigned int itemSubType,
+Container::Container(const unsigned int ID, const unsigned int itemType, const unsigned int itemSubType,
 					 const int maxUse, const int currUse) : Item(ID, itemType, itemSubType)
 {
 	this->zStacking = false;
