+++ Game/Source/Projectile.cpp
@@ -28,7 +28,7 @@ Projectile::Projectile(const Projectile* other)
 	this->zItemDescription = other->zItemDescription;
 }
 
-Projectile::Projectile(const long ID, const unsigned int itemType, const unsigned int itemSubType, 
+Projectile::Projectile(const unsigned int ID, const unsigned int itemType, const unsigned int itemSubType, 
 					   float velocity, float damage) : Item(ID, itemType, itemSubType)
 {
 	this->zStacking = true;
